# Conflicts:
#	build.gradle.kts
#	gradle.properties
#	src/main/java/net/frozenblock/wilderwild/block/PalmFrondsBlock.java
#	src/main/java/net/frozenblock/wilderwild/datagen/WWExperimentRecipeProvider.java
#	src/main/java/net/frozenblock/wilderwild/entity/AncientHornProjectile.java
#	src/main/java/net/frozenblock/wilderwild/entity/ai/warden/WardenNavigation.java

+++ src/main/java/net/frozenblock/wilderwild/block/PollenBlock.java
@@ -44,7 +44,7 @@ public class PollenBlock extends FlowerLichenBlock {
 				mutable.set(i + Mth.nextInt(random, -10, 10), j - random.nextInt(10), k + Mth.nextInt(random, -10, 10));
 				BlockState blockState = level.getBlockState(mutable);
 				if (!blockState.isCollisionShapeFullBlock(level, mutable) && !level.isRainingAt(mutable)) {
-					level.addParticle(RegisterParticles.POLLEN, (double) mutable.getX() + random.nextDouble(), (double) mutable.getY() + random.nextDouble(), (double) mutable.getZ() + random.nextDouble(), 0.0D, 0.0D, 0.0D);
+					level.addParticle(RegisterParticles.POLLEN, mutable.getX() + random.nextDouble(), mutable.getY() + random.nextDouble(), mutable.getZ() + random.nextDouble(), 0.0D, 0.0D, 0.0D);
 				}
 			}
 		}
